move main tests to their own file
